CM: Move PublicationState and ReviewWorkflowsStage into cell content
