+++ packages/core/admin/admin/src/content-manager/components/DynamicTable/CellContent/PublicationState/PublicationState.js
@@ -1,28 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import styled from 'styled-components';
 import { Status, Typography } from '@strapi/design-system';
 
-import { getTrad } from '../../../utils';
-
-const StyledStatus = styled(Status)`
-  width: min-content;
-`;
+import { getTrad } from '../../../../utils';
 
 export function PublicationState({ isPublished }) {
   const { formatMessage } = useIntl();
   const variant = isPublished ? 'success' : 'secondary';
 
   return (
-    <StyledStatus showBullet={false} variant={variant} size="S">
+    <Status showBullet={false} variant={variant} size="S" width="min-content">
       <Typography fontWeight="bold" textColor={`${variant}700`}>
         {formatMessage({
           id: getTrad(`containers.List.${isPublished ? 'published' : 'draft'}`),
           defaultMessage: isPublished ? 'Published' : 'Draft',
         })}
       </Typography>
-    </StyledStatus>
+    </Status>
   );
 }
